Move `type` key tests to schema test file
